requirements.txt
paho-mqtt==1.1
ipdb==0.10.1
ipdbplugin==1.4.5
nose==1.3.7
Flask==0.11.1
